chore: add SortTableHead
